requirements.txt
streamlit>=1.26.0
torch>=2.1.0
torchvision>=0.16.0
fastai>=2.7.13
Pillow>=10.0.0
replicate